Moved the delete, update and edit fucntions to ProductCatalogContext
